# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Reneo Backend API
  description: |
    Multi-seller commerce platform REST API.
    
    ## Authentication
    All protected endpoints require `Authorization: Bearer <token>` header.
    Obtain a token via `POST /auth/login`.
    
    ## Roles
    - **SELLER**: Can create/manage stores and products
    - **CUSTOMER**: Can browse products and place orders
    
    ## Money
    All prices are in whole FCFA (West African CFA Franc). No decimal subunit.
  version: 1.0.0
  contact:
    name: Chandradeep
    email: reneo-backend@chandradeep.dev
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT obtained from POST /auth/login
...
paths:
...
  /stores:
    post:
      tags: [Stores]
      summary: Create a store (seller only, one per seller)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name:
                  type: string
                  maxLength: 255
                slug:
                  type: string
                  pattern: '^[a-z0-9][a-z0-9-]*[a-z0-9]$'
                description:
                  type: string
                  maxLength: 5000
      responses:
        '201':
          description: Store created
        '409':
          description: Seller already has a store
  /stores/{slug}:
    get:
      tags: [Stores]
      summary: Get store profile with products
      security:
        - BearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Store profile with products
        '404':
          description: Store not found
...